display most recent messages (ugly code in user repository....) Bug where sending messages goes back from navigational view. need to fix
