+++ CPSC575Project/MessageScreen/MessageListVM.swift
@@ -16,9 +16,10 @@ class MessageListVM: ObservableObject{
     private var cancellables = Set<AnyCancellable>()
     
     init(){
-        userRepository.$matchedUsers.map { users in
-            users.map { user in
-                MessageCellVM(messageCell: MessagesCellModel(userId: user.uid!, name: user.name, image: user.imageUrl1, message: "need to somehow grab this"))
+        userRepository.$messageCellModel.map { msgs in
+            msgs.map { msg in
+                //I need to access chatRepository, use user uid to find most recent message and probably chat timestamp?
+                MessageCellVM(messageCell: msg)
             }
         }
         .assign(to: \.messageCellVM, on: self)
